# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ManagemenLogbook API",
        "description": "API documentation for ManagemenLogbook",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/projek": {
            "get": {
                "tags": [
                    "Projek"
                ],
                "summary": "List projects (Filter & Search)",
                "operationId": "993bb395ad6252d24c9181f76a8bac5f",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari nama projek",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "InProgress",
                                "Completed",
                                "OnHold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projek"
                ],
                "summary": "Create a project",
                "operationId": "ec98e2852bf38e7225b7de72394d9610",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "pic",
                                    "tgl_mulai",
                                    "tgl_selesai"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Sistem Managemen Logbook"
                                    },
                                    "pic": {
                                        "type": "string",
                                        "example": "Acid"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Projek untuk mengelola logbook tugas"
                                    },
                                    "tgl_mulai": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-10"
                                    },
                                    "tgl_selesai": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-30"
                                    }
                                },
                                "type": "object"
                            },
                            "example": "{\"nama\":\"Sistem Managemen Logbook\", \"pic\":\"Acid\",\"deskripsi\":\"Projek untuk mengelola logbook tugas\",\"tgl_mulai\":\"2026-01-10\",\"tgl_selesai\":\"2026-06-30\"}"
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projek created"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projek/{id}": {
            "get": {
                "tags": [
                    "Projek"
                ],
                "summary": "Get project details",
                "operationId": "a4aa8409e160739bb8618522bb64ff58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project details"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projek"
                ],
                "summary": "Update project",
                "operationId": "cfac919837ec0a4c32031192ff7261c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Nama Projek Baru"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Deskripsi proyek"
                                    },
                                    "tgl_mulai": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-10"
                                    },
                                    "tgl_selesai": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-30"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "InProgress"
                                    }
                                },
                                "type": "object"
                            },
                            "example": "{\"nama\":\"Nama Projek Baru\",\"deskripsi\":\"Deskripsi proyek\",\"tgl_mulai\":\"2026-01-10\",\"tgl_selesai\":\"2026-06-30\",\"status\":\"InProgress\"}"
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projek updated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projek"
                ],
                "summary": "Delete project",
                "operationId": "ca93048e4936fcb910831080474d3dfe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projek deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projek/{id}/stats": {
            "get": {
                "tags": [
                    "Projek"
                ],
                "summary": "Get Project Dashboard Stats",
                "description": "Mengambil statistik ringkas (Total Task, Completed, Progress %) via SP sp_get_dashboard_card_stats",
                "operationId": "837daa10858283849f1fc96f93c3f3d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Projek",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pjk_id": {
                                            "type": "integer"
                                        },
                                        "total_tasks": {
                                            "type": "integer",
                                            "example": 26
                                        },
                                        "completed_tasks": {
                                            "type": "integer",
                                            "example": 18
                                        },
                                        "project_progress": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 75.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projek tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projek/{id}/breakdown": {
            "get": {
                "tags": [
                    "Projek"
                ],
                "summary": "Get Project Breakdown (Modul & Kegiatan)",
                "description": "Mengambil detail progress per modul dan kegiatan via SP sp_get_project_breakdown",
                "operationId": "616cacc2f28c95c204715992bac5e14b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Projek",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "nama_item": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projek/{id}/recalculate": {
            "post": {
                "tags": [
                    "Projek"
                ],
                "summary": "Force Recalculate Progress",
                "operationId": "66f73b78a3a09708afbdc83b1655ddbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Projek",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projek/{id}/members": {
            "get": {
                "tags": [
                    "Projek"
                ],
                "summary": "Get list of project members",
                "description": "Mengambil daftar anggota yang terdaftar dalam projek tertentu",
                "operationId": "495cf04dc33551d4d5ce5e56c0a0c29a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Projek",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "usr_id": {
                                                        "type": "integer"
                                                    },
                                                    "usr_first_name": {
                                                        "type": "string"
                                                    },
                                                    "usr_last_name": {
                                                        "type": "string"
                                                    },
                                                    "mpk_role_projek": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projek tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}